+++ sdk/c4-client.ts
@@ -2,8 +2,8 @@
 // Generated from /openapi.json via src/openapi-sdk-gen.js.
 // Do not edit by hand — re-run `c4 openapi --sdk` to refresh.
 
-// Spec version: 1.10.93
-// Generated at: 2026-05-02T07:17:13.816Z
+// Spec version: 1.10.94
+// Generated at: 2026-05-02T07:25:47.933Z
 
 export interface postAuthLoginBody {
   user: string; /** Username */